+++ src/shared/hooks/useBottomSheetDrag.ts
@@ -1,49 +1,81 @@
-import { useRef, useState, useCallback, useEffect } from 'react';
+import { useCallback, useRef, type RefObject } from 'react';
 
-import { SHEET_DURATION_MS } from '../constants/bottomSheet';
+import { SHEET_DURATION_MS } from '@/shared/constants/bottomSheet';
 
 interface UseBottomSheetDragProps {
+  sheetRef: RefObject<HTMLDivElement | null>;
+  threshold: number;
   onClose: () => void;
-  threshold?: number; // 드래그로 닫히는 임계치(px)
 }
 
-export const useBottomSheetDrag = ({
-  onClose,
-  threshold = 100,
-}: UseBottomSheetDragProps) => {
-  const sheetRef = useRef<HTMLDivElement | null>(null);
-  const [isDragging, setIsDragging] = useState(false); // 드래그 중 여부 (UI용)
-  const isDraggingRef = useRef(false); // 리스너 가드용
+interface UseBottomSheetDragReturn {
+  onHandlePointerDown: (e: React.PointerEvent) => void;
+  animateClose: () => void;
+}
 
-  // 닫기 애니메이션 (시트 변경 후 onClose 호출)
+/**
+ * Pointer Event 기반 드래그 (마우스, 터치, 펜 지원)
+ * 바텀시트 드래그/닫기 동작을 담당
+ * backdrop 닫기에도 동일 애니메이션을 재사용할 수 있도록 animateClose 제공
+ */
+export const useBottomSheetDrag = ({
+  sheetRef,
+  threshold,
+  onClose,
+}: UseBottomSheetDragProps): UseBottomSheetDragReturn => {
+  const isDraggingRef = useRef(false);
+
+  // 드래그로 닫을 때
   const animateClose = useCallback(() => {
     if (!sheetRef.current) {
       onClose();
       return;
     }
-    const el = sheetRef.current;
-    el.style.transition = `transform ${SHEET_DURATION_MS}ms ease-in-out`;
-    el.style.transform = 'translate(-50%, 100%)';
+    const sheet = sheetRef.current;
+    // 하단으로 이동 후 onClose 호출
+    sheet.style.transition = `transform ${SHEET_DURATION_MS}ms ease-in-out`;
+    sheet.style.transform = 'translate(-50%, 100%)';
 
-    window.setTimeout(() => {
-      el.style.transition = '';
-      el.style.transform = '';
+    setTimeout(() => {
       onClose();
-    }, 300);
-  }, [onClose]);
+    }, SHEET_DURATION_MS);
+  }, [onClose, sheetRef]);
 
-  // Pointer Event 기반 드래그 (마우스, 터치, 펜 모두 처리)
-  const handlePointerDown = useCallback(
-    (e: React.PointerEvent<HTMLDivElement>) => {
+  const onHandlePointerDown = useCallback(
+    (e: React.PointerEvent) => {
       e.preventDefault();
       e.stopPropagation();
 
-      setIsDragging(true);
+      // 드래그 시작 상태 설정 (useRef로 즉시 반영)
       isDraggingRef.current = true;
       const startY = e.clientY;
+      const target = e.currentTarget as HTMLElement;
 
+      // 포인터 캡처 설정 (윈도우 밖 이동 시에도 이벤트 수신)
+      target.setPointerCapture?.(e.pointerId);
+
+      // 이벤트 리스너 정리 함수
+      const cleanup = () => {
+        target.releasePointerCapture?.(e.pointerId);
+        target.removeEventListener('pointermove', handlePointerMove);
+        target.removeEventListener('pointerup', handlePointerUp);
+        target.removeEventListener('pointercancel', handlePointerCancel);
+        target.removeEventListener('lostpointercapture', handlePointerCancel);
+      };
+
+      // 포인터 취소/캡처 해제 시 정리
+      const handlePointerCancel = () => {
+        isDraggingRef.current = false;
+        // 원위치로 복귀
+        if (sheetRef.current) {
+          sheetRef.current.style.transition = `transform ${SHEET_DURATION_MS}ms ease-in-out`;
+          sheetRef.current.style.transform = 'translate(-50%, 0)';
+        }
+        cleanup();
+      };
+
+      // 드래그 중 실시간 움직임 처리
       const handlePointerMove = (ev: PointerEvent) => {
-        if (!isDragging) return;
         if (!isDraggingRef.current) return;
 
         ev.preventDefault();
@@ -52,51 +84,49 @@ export const useBottomSheetDrag = ({
         const deltaY = ev.clientY - startY;
 
         if (deltaY > 0 && sheetRef.current) {
+          // 드래그 거리를 threshold 내에서 제한하여 자연스러운 움직임 구현
           const translateY = Math.min(deltaY, threshold);
           sheetRef.current.style.transform = `translate(-50%, ${translateY}px)`;
+          // 드래그 중에는 transition을 비활성화
           sheetRef.current.style.transition = 'none';
         }
       };
 
+      // 드래그 종료 시 동작 결정 (닫기 또는 원위치 복귀)
       const handlePointerUp = (ev: PointerEvent) => {
-        setIsDragging(false);
         isDraggingRef.current = false;
         const deltaY = ev.clientY - startY;
 
         if (sheetRef.current) {
           if (deltaY > threshold) {
+            // threshold를 넘으면 닫기
             animateClose();
           } else {
-            sheetRef.current.style.transform = 'translate(-50%, 0)';
+            // threshold 미만이면 원위치로 복귀
             sheetRef.current.style.transition = `transform ${SHEET_DURATION_MS}ms ease-in-out`;
+            sheetRef.current.style.transform = 'translate(-50%, 0)';
           }
         }
 
-        document.removeEventListener('pointermove', handlePointerMove);
-        document.removeEventListener('pointerup', handlePointerUp);
+        cleanup();
       };
 
-      document.addEventListener('pointermove', handlePointerMove);
-      document.addEventListener('pointerup', handlePointerUp);
+      // 대상 요소에 이벤트 리스너 등록
+      target.addEventListener('pointermove', handlePointerMove, {
+        passive: false,
+      });
+      target.addEventListener('pointerup', handlePointerUp, {
+        passive: false,
+      });
+      target.addEventListener('pointercancel', handlePointerCancel, {
+        passive: false,
+      });
+      target.addEventListener('lostpointercapture', handlePointerCancel, {
+        passive: false,
+      });
     },
-    [threshold, animateClose, isDragging]
+    [sheetRef, threshold, animateClose]
   );
 
-  // 정리
-  useEffect(() => {
-    const el = sheetRef.current;
-    return () => {
-      if (el) {
-        el.style.transition = '';
-        el.style.transform = '';
-      }
-    };
-  }, []);
-
-  return {
-    sheetRef,
-    isDragging,
-    handlePointerDown,
-    closeWithAnimation: animateClose,
-  };
+  return { onHandlePointerDown, animateClose };
 };
